unneeded input removed and other fixes

+++ OOP/IFuelConsummable.cs
@@ -6,6 +6,6 @@ namespace OOP
 {
     public interface IFuelConsummable
     {
-        int FuelConsumed();
+        double FuelConsumed();
     }
 }
